refactor: move ShareArtifactsContainer out of ShareView
